+++ changelog.md
@@ -15,6 +15,7 @@
 - [x] f64 printing
 - [x] Verify only compilation mode
 - [x] Expression chaining `(1 + 2 + 3)`
+- [x] Opperator Precedence
 
 ### Fixes
 - [x] Not parsing structures properly on return